feat: Created and filled Seeder Game

+++ Final-Proyect-Backend/database/migrations/2023_04_13_154003_create_games_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('game_image')->nullable(true);
             $table->string('name')->unique()->nullable(false);
             $table->string('description')->nullable(false);
-            $table->integer('score')->nullable(true);
+            $table->float('score')->nullable(true);
             $table->string('genre')->nullable(false);
             $table->string('publisher')->nullable(false);
             $table->date('release_date')->nullable(false);
